Move Zarr-family parsers into a parsers/zarr/ package
ZarrParser and ZippedZarrParser are closely related — both map on-disk
Zarr stores to ManifestStores and share metadata/manifest-building
machinery — so group them in a parsers/zarr/ directory:

- zarr/zarr.py: ZarrParser and its obstore-listing chunk discovery
- zarr/zip.py: ZippedZarrParser and the zip central-directory machinery
- zarr/common.py: the shared helpers (ZarrFormat, metadata_as_v3,
  parse_array_layout, chunk_entries_to_manifest, join_url, _run_async)

The package __init__ re-exports everything previously importable from
virtualizarr.parsers.zarr, so external imports are unaffected.
