+++ virtualizarr/parsers/zarr/__init__.py
@@ -1,0 +1,7 @@
+from virtualizarr.parsers.zarr.zarr import ZarrParser
+from virtualizarr.parsers.zarr.zip import ZippedZarrParser
+
+__all__ = [
+    "ZarrParser",
+    "ZippedZarrParser",
+]
